+++ frontend/retailRedStorefront.js
@@ -1,4 +1,6 @@
-import { config } from './config';
+import getConfig from './helpers/getConfig';
+
+const { config } = getConfig();
 
 let retailred;
 
@@ -19,40 +21,13 @@ export default () => {
       showExactUntil: config.inventoryShowExactUntil,
       showLowUntil: config.inventoryShowLowUntil,
     },
+    localization: config.localization,
     legal: {
       terms: config.termsLink,
       privacy: config.privacyLink,
     },
-    // TODO: remove
-    product: {
-      code: 'acai-bowl',
-      name: 'Dark Blue Dress',
-      options: [
-        {
-          code: 'code-color',
-          name: 'Color',
-          value: {
-            code: 'code-color-red',
-            name: 'Red',
-          },
-        },
-        {
-          code: 'code-size',
-          name: 'Size',
-          value: {
-            code: 'code-size-m',
-            name: 'M',
-          },
-        },
-      ],
-      quantity: 2,
-      imageUrl: 'https://demo.elliselaine.com/pub/media/catalog/product/cache/5e62a3e057e3c5cc5f2bb9763ad693f7/k/l/kleid_dunkelblau_5.jpg',
-      price: 34.0,
-      currencyCode: 'USD',
-      identifiers: {
-        sku: 'abc123',
-      },
-    },
+    newsletterOptIn: config.newsletterOptIn,
+    platform: "engage"
   });
   return retailred;
 };