+++ frontend/selectors.js
@@ -1,17 +1,13 @@
 import { createSelector } from 'reselect';
-import { getUserDisplayName, getUserEmail } from '@shopgate/engage/user';
+import { getProduct } from '@shopgate/engage/product'
+import { formatProductData } from './helpers/formatProductData';
 
-export const getUserData = createSelector(
-  getUserDisplayName,
-  getUserEmail,
-  (name, email) => {
-    if (!email || !name) {
-      return null;
-    }
-
-    return {
-      name,
-      email,
-    };
-  }
+/**
+ * Gets the current product in a formatted way.
+ * @param {Object} state The current state.
+ * @returns {Object} The formatted selected variant.
+ */
+export const getProductFormatted = createSelector(
+  getProduct,
+  formatProductData
 );