+++ frontend/subscriptions.js
@@ -1,10 +1,17 @@
 import { appDidStart$ } from '@shopgate/engage/core';
-import { productIsReady$ } from '@shopgate/pwa-tracking/streams/product';
+import { productIsReady$ } from '../../../libraries/tracking/streams/product';
 import { variantDidChange$ } from '@shopgate/pwa-common-commerce/product/streams';
+import {
+  userDataReceived$,
+  userDidLogout$,
+  getUserData
+} from '@shopgate/engage/user';
+import { getProductFormatted } from './selectors';
 import getRetailred from './retailRedStorefront';
 
 export default (subscribe) => {
-  subscribe(appDidStart$, ({ getState }) => {
+  // Include storefront library
+  subscribe(appDidStart$, () => {
     const script = document.createElement('script');
     script.type = 'text/javascript';
     script.async = true;
@@ -13,27 +20,66 @@ export default (subscribe) => {
     parent.appendChild(script);
   });
 
-  // TODO: set user data if already logged in
-  // TODO: call updateConfig() if user did login/logout
-  // TODO: disable if product is not orderable
+  // Update user data with logged in user
+  subscribe(userDataReceived$, ({ getState }) => {
+    const state = getState();
+    const userData = getUserData(state);
 
-  subscribe(productIsReady$, ({ getState }) => {
+    const user = userData ? {
+        code: null,
+        firstName: userData.firstName,
+        lastName: userData.lastName,
+        phone: userData.phone,
+        emailAddress: userData.mail
+      } : {}
 
     getRetailred().updateConfig({
-      product: {}, // TODO: pass data here
+      customer: user,
     });
   });
 
-  subscribe(variantDidChange$, ({ getState, action }) => {
+  // Unset user after logout
+  subscribe(userDidLogout$, () => {
+    const user = {
+      code: null,
+      firstName: '',
+      lastName: '',
+      phone: '',
+      emailAddress: ''
+    }
+    getRetailred().updateConfig({
+      customer: user,
+    });
+  });
+
+  // Update product data
+  subscribe(productIsReady$, ({ getState, action }) => {
     const state = getState();
-    const { id: productId } = action.productData;
+    const {
+      id: productId,
+    } = action.productData;
     const props = { productId };
 
-    getRetailred().updateConfig({
-      product: {}, // TODO: pass data here
-    });
+    const formatedProduct = getProductFormatted(state, props)
 
+    getRetailred().updateConfig({
+      product: formatedProduct,
+    });
+  });
+
+  // Update product data after variant change
+  subscribe(variantDidChange$, ({ getState, action }) => {
+    const state = getState();
+    const {
+      id: productId,
+    } = action.productData;
+    const props = { productId };
+
+    const formatedProduct = getProductFormatted(state, props)
+
+    getRetailred().updateConfig({
+      product: formatedProduct
+    });
   });
 
 };
-